admin-back/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
fastapi==0.123.10
h11==0.16.0
idna==3.10
pydantic==2.11.0a1
pydantic_core==2.28.0
python-jose[cryptography]==3.4.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.49.1
typing_extensions==4.12.2
uvicorn==0.34.0
